bytearray_stubs.c: avoid deprecated bigarray names
In 2006 the Bigarray API was rewritten to consistently use the caml_
namespacing prefix, but the old definitions were left in place for
backward-compatibility.
  https://github.com/ocaml/ocaml/commit/f63965c1e4a5d4e41e27

The old definitions are being actively deprecated during the
transition to OCaml 5, so the parmap build now prints deprecation
warnings.
  https://github.com/ocaml/ocaml/commit/f6190f3d0c49c5220d44

The present commit migrates the Parmap codebase to use the new names,
removing all the warnings on the C stubs.